Complete test for 'subclass_of' predicate factory
